extract matchSpecApiByName logic
