+++ protocol/chainlib/grpc.go
@@ -6,7 +6,6 @@ import (
 	"fmt"
 	"net"
 	"net/http"
-	"regexp"
 	"sync"
 	"time"
 
@@ -17,10 +16,6 @@ import (
 	"github.com/lavanet/lava/utils"
 	pairingtypes "github.com/lavanet/lava/x/pairing/types"
 	spectypes "github.com/lavanet/lava/x/spec/types"
-)
-
-const (
-	grpcInterface = "grpc"
 )
 
 type GrpcChainParser struct {
@@ -75,7 +70,7 @@ func (apip *GrpcChainParser) getSupportedApi(name string) (*spectypes.ServiceApi
 	defer apip.rwLock.RUnlock()
 
 	// Fetch server api by name
-	api, ok := apip.MatchSpecApiByName(name)
+	api, ok := matchSpecApiByName(name, apip.serverApis)
 
 	// Return an error if spec does not exist
 	if !ok {
@@ -88,30 +83,6 @@ func (apip *GrpcChainParser) getSupportedApi(name string) (*spectypes.ServiceApi
 	}
 
 	return &api, nil
-}
-
-func (apip *GrpcChainParser) MatchSpecApiByName(name string) (spectypes.ServiceApi, bool) {
-	// Guard that the JsonRPCChainParser instance exists
-	if apip == nil {
-		return spectypes.ServiceApi{}, false
-	}
-
-	// Acquire read lock
-	apip.rwLock.RLock()
-	defer apip.rwLock.RUnlock()
-
-	// TODO: make it faster and better by not doing a regex instead using a better algorithm
-	for apiName, api := range apip.serverApis {
-		re, err := regexp.Compile(apiName)
-		if err != nil {
-			utils.LavaFormatError("regex Compile api", err, &map[string]string{"apiName": apiName})
-			continue
-		}
-		if re.Match([]byte(name)) {
-			return api, true
-		}
-	}
-	return spectypes.ServiceApi{}, false
 }
 
 // SetSpec sets the spec for the GrpcChainParser
